Revert "change to weak level default printers #140 #77 (#143)"
This reverts commit 340515bca6bcc9dc65ec9d765944c4c61159c988.
