+++ include/internal/iutest_option_message.hpp
@@ -195,7 +195,6 @@ inline void iuOptionMessage::ShowSpec()
     IIUT_SHOW_MACRO(IUTEST_HAS_CXX_HDR_STRING_VIEW);
     IIUT_SHOW_MACRO(IUTEST_HAS_CXX_HDR_VARIANT);
     IIUT_SHOW_MACRO(IUTEST_HAS_CXX11);
-    IIUT_SHOW_MACRO(IUTEST_HAS_CXX1Z);
     IIUT_SHOW_MACRO(IUTEST_HAS_DECLTYPE);
     IIUT_SHOW_MACRO(IUTEST_HAS_DELETED_FUNCTIONS);
     IIUT_SHOW_MACRO(IUTEST_HAS_EXCEPTIONS);
