+++ include/iutest_assertion.hpp
@@ -17,7 +17,6 @@
 
 //======================================================================
 // include
-#include "iutest_any.hpp"
 #include "iutest_result.hpp"
 #include "iutest_printers.hpp"
 #include "internal/iutest_list.hpp"
