+++ include/iutest_param_tests.hpp
@@ -15,6 +15,10 @@
 #ifndef INCG_IRIS_IUTEST_PARAM_TESTS_HPP_3DBBF55E_A076_4207_B49E_34734EE333EE_
 #define INCG_IRIS_IUTEST_PARAM_TESTS_HPP_3DBBF55E_A076_4207_B49E_34734EE333EE_
 
+//======================================================================
+// include
+#include "iutest_any.hpp"
+
 #if IUTEST_HAS_PARAM_TEST
 
 //======================================================================
